+++ src/main/java/online/kingdomkeys/kingdomkeys/client/gui/elements/buttons/MenuSelectMagicButton.java
@@ -48,7 +48,7 @@ public class MenuSelectMagicButton extends MenuButtonBase {
 
 					if (slot <= MenuMagicSelectorScreen.BAG_OFFSET) {
 						int bagSlot = Math.abs(slot - MenuMagicSelectorScreen.BAG_OFFSET);
-						if (!Utils.hasOnlyOneBag(player, BagItem.Type.MAGICS_BAG)) //Only one bag should be in the inv
+						if (!Utils.hasOnlyOneBag(player, BagItem.Type.SPELLS_BAG)) //Only one bag should be in the inv
 							return;
 
 						ItemStack magicBag = player.getInventory().getItem(Utils.getMagicBagSlot(player));