+++ src/main/java/online/kingdomkeys/kingdomkeys/client/gui/menu/items/MeldingScreen.java
@@ -206,7 +206,7 @@ public class MeldingScreen extends MenuFilterable {
 			entries.add(new SlotEntry(slot, stack.copy(), false, false));
 		}
 
-		if (Utils.hasOnlyOneBag(player, BagItem.Type.MAGICS_BAG)) {
+		if (Utils.hasOnlyOneBag(player, BagItem.Type.SPELLS_BAG)) {
 			ItemStack magicBag = ItemStack.EMPTY;
 
 			for (ItemStack stack : minecraft.player.getInventory().items) {