+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 ## Added:
 * Minimap icons for Air Soldier, Defender, Neoshadow and Novashadow (Thanks to WillabyNeko)
 * Level up item giving tracker so further exp readjustments won't give them again.
+* Keychains bag.
+
+## Changed:
+* Pyramid gummi hitbox got adjusted.
 
 ## Fixed:
 * Grinding on rails not showing to remote players the flowmotion trail.
 * Antiform revert sound still being the unsummon one.
+* Keychain equipment screen button width overlapping the scroll bar.
 * [EFM] Keyblades in dual wield boxing animation, removed their keychain requirement, now they will work as long as the held items are keyblades.
 * [EFM] idle animations on some drive forms.
 * [EFM] Limit form boxing instead of attacking with the keyblade.